pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "chess-vision"
version = "0.1.0"
description = "CNN-powered chess board recognition from a USB camera. Records OTB games for engine analysis."
requires-python = ">=3.11"
dependencies = [
    "torch>=2.0",
    "torchvision>=0.15",
    "onnx>=1.14",
    "onnxruntime>=1.16",
    "opencv-python>=4.8",
    "python-chess>=1.10",
    "click>=8.1",
    "pyyaml>=6.0",
    "numpy>=1.24",
    "Pillow>=10.0",
    "scikit-learn>=1.3",
    "tqdm>=4.66",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
]

[project.scripts]
chess-vision = "chess_vision.cli:main"

[tool.setuptools.packages.find]
include = ["chess_vision*"]
